Add LinkedListUtils file
