+++ src/main/java/com/company/linkedlist/LinkedListUtils.java
@@ -1,0 +1,14 @@
+package com.company.linkedlist;
+
+public class LinkedListUtils {
+    public static boolean areEqual(ListNode head1, ListNode head2) {
+        while (head1 != null && head2 != null) {
+            if (head1.val != head2.val) {
+                return false;
+            }
+            head1 = head1.next;
+            head2 = head2.next;
+        }
+        return head1 == null && head2 == null;
+    }
+}
