+++ src/main/java/com/company/linkedlist/ListNode.java
@@ -1,29 +1,15 @@
 package com.company.linkedlist;
 
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
 public class ListNode {
     public int val;
     public ListNode next;
 
-    public ListNode() {
-    }
-
     public ListNode(int val) {
         this.val = val;
-    }
-
-    public ListNode(int val, ListNode next) {
-        this.val = val;
-        this.next = next;
-    }
-
-    public static boolean areEqual(ListNode head1, ListNode head2) {
-        while (head1 != null && head2 != null) {
-            if (head1.val != head2.val) {
-                return false;
-            }
-            head1 = head1.next;
-            head2 = head2.next;
-        }
-        return head1 == null && head2 == null;
     }
 }
